make the config module pluggable.
In some case it could be interersted to change the config module and replace it
by another. useful for example if you're storing your conf in a distributed
database or such things.

While I'm handle all the config changes in the `coffer_config` module so the
http modules just have to call these methods. There is a special case for the
http changes to handle the restart.